Issue 44564: Clean up folder import
