+++ api/schemas/folder.xsd
@@ -81,7 +81,7 @@
                 <xs:element name="queries" minOccurs="0" type="exportDirType" />
                 <xs:element name="views" minOccurs="0" type="exportDirType" />
                 <xs:element name="lists" minOccurs="0" type="exportDirType" />
-                <xs:element name="missingValueIndicators" type="study:missingValueIndicatorsType" minOccurs="0" />
+                <xs:element name="missingValueIndicators" type="missingValueIndicatorsType" minOccurs="0" />
                 <xs:element name="wikis" minOccurs="0" type="exportDirType"/>
                 <xs:element name="etls" minOccurs="0" type="exportDirType"/>
                 <xs:element name="externalSchemas" minOccurs="0" type="exportDirType"/>
@@ -253,6 +253,37 @@
         <xs:attribute name="value" type="xs:string">
             <xs:annotation>
                 <xs:documentation>Container-specific value of the property</xs:documentation>
+            </xs:annotation>
+        </xs:attribute>
+    </xs:complexType>
+
+    <xs:complexType name="missingValueIndicatorsType">
+        <xs:sequence minOccurs="0" maxOccurs="unbounded">
+            <xs:element name="missingValueIndicator">
+                <xs:complexType>
+                    <xs:attribute name="indicator" type="xs:string">
+                        <xs:annotation>
+                            <xs:documentation>
+                                The indicator to use for a certain type of missing values (e.g., "N").
+                            </xs:documentation>
+                        </xs:annotation>
+                    </xs:attribute>
+                    <xs:attribute name="label" type="xs:string">
+                        <xs:annotation>
+                            <xs:documentation>
+                                The text to use in association with this indicator in the UI
+                                (e.g., "Required field marked by site as 'data not available'.").
+                            </xs:documentation>
+                        </xs:annotation>
+                    </xs:attribute>
+                </xs:complexType>
+            </xs:element>
+        </xs:sequence>
+        <xs:attribute name="file" type="xs:string">
+            <xs:annotation>
+                <xs:documentation>
+                    The file that contains the list of missing value indicators.
+                </xs:documentation>
             </xs:annotation>
         </xs:attribute>
     </xs:complexType>
