+++ api/schemas/study.xsd
@@ -238,7 +238,6 @@
 				<xs:element name="queries" minOccurs="0" type="exportDirType" />
 				<xs:element name="views" minOccurs="0" type="exportDirType" />
                 <xs:element name="lists" minOccurs="0" type="exportDirType" />
-				<xs:element name="missingValueIndicators" type="missingValueIndicatorsType" minOccurs="0" />
                 <xs:element name="comments" minOccurs="0">
                     <xs:complexType>
                         <xs:attribute name="participantCommentDatasetId" type="xs:int">
@@ -503,37 +502,6 @@
             <xs:annotation>
                 <xs:documentation>
                       Names the directory that contains exported resource.
-                </xs:documentation>
-            </xs:annotation>
-        </xs:attribute>
-    </xs:complexType>
-
-    <xs:complexType name="missingValueIndicatorsType">
-        <xs:sequence minOccurs="0" maxOccurs="unbounded">
-            <xs:element name="missingValueIndicator">
-                <xs:complexType>
-                    <xs:attribute name="indicator" type="xs:string">
-                        <xs:annotation>
-                            <xs:documentation>
-                                The indicator to use for a certain type of missing values (e.g., "N").
-                            </xs:documentation>
-                        </xs:annotation>
-                    </xs:attribute>
-                    <xs:attribute name="label" type="xs:string">
-                        <xs:annotation>
-                            <xs:documentation>
-                                The text to use in association with this indicator in the UI
-                                (e.g., "Required field marked by site as 'data not available'.").
-                            </xs:documentation>
-                        </xs:annotation>
-                    </xs:attribute>
-                </xs:complexType>
-            </xs:element>
-        </xs:sequence>
-        <xs:attribute name="file" type="xs:string">
-            <xs:annotation>
-                <xs:documentation>
-                      The file that contains the list of missing value indicators.
                 </xs:documentation>
             </xs:annotation>
         </xs:attribute>
