+++ study/src/org/labkey/study/importer/CreateChildStudyPipelineJob.java
@@ -410,8 +410,7 @@ public class CreateChildStudyPipelineJob extends AbstractStudyPipelineJob
         {
             importContext = new StudyImportContext(getUser(), newStudy.getContainer(), studyDoc, null, new PipelineJobLoggerGetter(this), studyDir);
 
-            // missing values and qc states
-            new MissingValueImporterFactory().create().process(null, importContext, studyDir);
+            // old study-based qc states: should remove this at some point
             new StudyQcStatesImporter().process(importContext, studyDir, errors);
 
             // dataset definitions
